mp3.2

+++ student-distrib/tests.c
@@ -8,7 +8,7 @@
 #define FAIL 0
 
 extern void rtc_init();
-extern int* filesys_base;
+extern int8_t* filesys_base;
 /* format these macros as you see fit */
 #define TEST_HEADER 	\
 	printf("[TEST %s] Running %s at %s:%d\n", __FUNCTION__, __FUNCTION__, __FILE__, __LINE__)
